+++ src/tcerror.h
@@ -111,6 +111,8 @@ DEFINE_TCError(CastOp, "unsupported cast op: `%s' type -> `%s' type");
 DEFINE_TCError(UnmatchParam, "not match parameter, require size is %d, but is %d");
 DEFINE_TCError(RegexSyntax, "regex syntax error: `%s'");
 DEFINE_TCError(NoCommonSuper, "cannot resolve common super type from `%s'");
+DEFINE_TCError(ConcatParam,
+               "concatenation of `%s' type value is not allowed in parameter expansion");
 
 #undef DEFINE_TCError