+++ CHANGELOG.md
@@ -85,6 +85,7 @@
     - module private member access
     - read-only symbol/field access
     - no-return expression checking in finally-block
+    - illegal concatenation of parameter expansion
 - preserve exit status during the evaluation of finally-block
 - complete type template, such as ``Array``, ``Map``, ``Tuple``, ``Option``
 - do not complete methods that do not satisfy type constraints